Removed relative imports as requested. Moved core tests into main tests subdirectory to be consistent with other tests.
