--- src/psyclone/core/tests/conftest.py
@@ -1,63 +1,0 @@
-# -----------------------------------------------------------------------------
-# BSD 3-Clause License
-#
-# Copyright (c) 2017-2019, Science and Technology Facilities Council.
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-# * Redistributions of source code must retain the above copyright notice, this
-#   list of conditions and the following disclaimer.
-#
-# * Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-#
-# * Neither the name of the copyright holder nor the names of its
-#   contributors may be used to endorse or promote products derived from
-#   this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-# COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
-# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
-# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-# -----------------------------------------------------------------------------
-# Author A. R. Porter, STFC Daresbury Lab
-
-
-''' Module which performs pytest set-up so that we can specify
-    command-line options '''
-
-from __future__ import absolute_import
-import pytest
-
-
-@pytest.fixture(scope="session", autouse=True)
-def setup_psyclone_config():
-    '''This per session fixture defines the environment variable
-    PSYCLONE_CONFIG to point to the config file included in the
-    PSyclone repo. This way all tests will get the same config,
-    independent of a potential psyclone config file installed by
-    the user.
-    '''
-    from psyclone.configuration import Config
-    import os
-    config_file = Config.get_repository_config_file()
-
-    # In case that PSyclone is installed and tested (e.g. travis),
-    # the 'repository' config file does not exist (since it is
-    # installed in a different directory). In that case the standard
-    # search path of the Configuration object will find the right
-    # config file. So only overwrite the default search path behaviour
-    # if the repository config file is actually found:
-    if os.path.isfile(config_file):
-        os.environ["PSYCLONE_CONFIG"] = config_file
