+++ src/psyclone/tests/dynamo0p3_basis_test.py
@@ -47,8 +47,8 @@ from psyclone.parse.algorithm import parse
 from psyclone.parse.utils import ParseError
 from psyclone.psyGen import PSyFactory, GenerationError, InternalError
 from psyclone.dynamo0p3 import DynKernMetadata, DynKern
-from .dynamo0p3_build import Dynamo0p3Build
-from .psyclone_test_utils import print_diffs
+from psyclone.tests.dynamo0p3_build import Dynamo0p3Build
+from psyclone.tests.utilities import print_diffs
 
 # constants
 BASE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
