+++ src/psyclone/tests/dynamo0p3_transformations_test.py
@@ -56,7 +56,7 @@ from psyclone.transformations import TransformationError, \
     Dynamo0p3KernelConstTrans
 from psyclone.configuration import Config
 
-from .dynamo0p3_build import Dynamo0p3Build
+from psyclone.tests.dynamo0p3_build import Dynamo0p3Build
 
 
 # The version of the API that the tests in this file
