+++ src/psyclone/tests/gocean1p0_stencil_test.py
@@ -48,7 +48,7 @@ from psyclone.parse.utils import ParseError
 from psyclone.gocean1p0 import GOStencil
 from psyclone import expression as expr
 
-from .gocean1p0_build import GOcean1p0Build
+from psyclone.tests.gocean1p0_build import GOcean1p0Build
 
 API = "gocean1.0"
 BASE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
