+++ src/psyclone/tests/kernel_transformation_test.py
@@ -48,8 +48,8 @@ from psyclone.psyGen import Kern
 from psyclone.generator import GenerationError
 from psyclone.configuration import Config
 
-from .dynamo0p3_build import Dynamo0p3Build
-from .psyclone_test_utils import get_invoke
+from psyclone.tests.dynamo0p3_build import Dynamo0p3Build
+from psyclone.tests.utilities import get_invoke
 
 
 def setup_module():
@@ -192,7 +192,7 @@ def test_new_kernel_file(kernel_outputdir, monkeypatch):
             break
     assert found
 
-    from .gocean1p0_build import GOcean1p0Build
+    from psyclone.tests.gocean1p0_build import GOcean1p0Build
     # If compilation fails this will raise an exception
     GOcean1p0Build(kernel_outputdir).compile_file(filename)
