+++ src/psyclone/tests/psyir/backend/fortran_test.py
@@ -44,7 +44,7 @@ from psyclone.psyir.backend.base import VisitorError
 from psyclone.psyir.backend.fortran import gen_intent, gen_dims, gen_kind, \
     FortranWriter
 from psyclone.psyGen import Symbol, Fparser2ASTProcessor, Node, CodeBlock
-from ...psyclone_test_utils import create_schedule
+from psyclone.tests.utilities import create_schedule
 
 
 @pytest.fixture(scope="function")
